+++ src/__tests__/foliage.test.js
@@ -41,6 +41,7 @@ describe('Foliage', function() {
       let query = plant.get('first')
 
       query.set('second', 'modified')
+
       plant.get([ 'first', 'second' ]).valueOf().should.equal('modified')
     })
   })
@@ -69,7 +70,6 @@ describe('Foliage', function() {
 
     it ('map over top level elements', function() {
       let plant = new Foliage([ 1, 2, 3])
-
       plant.map(incr).should.eql([ 2, 3, 4])
     })
 
@@ -143,6 +143,21 @@ describe('Foliage', function() {
       let query = plant.get('first')
 
       query.trunk().should.equal(plant)
+    })
+  })
+
+  describe('Foliage::toJSON', function() {
+    let data = [1,2,3,4]
+    it ('returns the source value of it is root', function() {
+      let plant = new Foliage(data)
+      plant.toJSON().should.equal(data)
+    })
+
+    it ('returns the subset if it is a cursor', function() {
+      let plant = new Foliage({ first: data })
+      let query = plant.get('first')
+
+      query.toJSON().should.equal(data)
     })
   })