+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0
+
+- Add toJSON methods
+- Removed internal Cursor class. Foliage all the way down...
+
 ## 0.1.0
 
 - Broke out dedicated `Cursor` and `functions` modules.